backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
sqlalchemy
pydantic
python-dotenv
httpx
jinja2
psycopg2-binary